+++ lib/constructs/StreamlitEcrStackProps.ts
